moved the base serializer, factory and built in serializables in pymodaq_utils. data serialization is declared during object instantiation in pymodaq_data. Same for Parameter.
